requirements.txt
asgiref==3.8.1
astroid==3.1.0
certifi==2024.2.2
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.7.1
dill==0.3.8
Django==5.0.4
djangorestframework==3.15.1
drf-yasg==1.21.7
httpie==3.2.2
idna==3.7
inflection==0.5.1
isort==5.13.2
itypes==1.2.0
Jinja2==3.1.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mccabe==0.7.0
mdurl==0.1.2
multidict==6.0.5
packaging==24.0
platformdirs==4.2.0
psycopg2-binary==2.9.9
Pygments==2.17.2
pylint==3.1.0
pylint-django==2.5.5
pylint-plugin-utils==0.8.2
PySocks==1.7.1
python-decouple==3.8
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.7.1
sqlparse==0.5.0
tomli==2.0.1
tomlkit==0.12.4
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==2.2.1
